Expose validateTs and compileTs
