+++ packages/plugins/typescript/apollo-angular/tests/apollo-angular.spec.ts
@@ -1,11 +1,10 @@
-import '@graphql-codegen/testing';
+import { validateTs } from '@graphql-codegen/testing';
 import gql from 'graphql-tag';
 import { plugin, addToSchema } from '../src/index';
 import { parse, GraphQLSchema, buildClientSchema, buildSchema, extendSchema } from 'graphql';
 import { Types, mergeOutputs } from '@graphql-codegen/plugin-helpers';
 import { plugin as tsPlugin } from '../../typescript/src/index';
 import { plugin as tsDocumentsPlugin } from '../../../typescript/operations/src/index';
-import { validateTs } from '../../typescript/tests/validate';
 import { readFileSync } from 'fs';
 
 describe('Apollo Angular', () => {
