+++ packages/plugins/typescript/mongodb/tests/typescript-mongo.spec.ts
@@ -1,8 +1,7 @@
-import '@graphql-codegen/testing';
+import { validateTs } from '@graphql-codegen/testing';
 import { plugin, addToSchema } from './../src/index';
 import { buildSchema, print, GraphQLSchema } from 'graphql';
 import { plugin as tsPlugin } from '../../typescript/src/index';
-import { validateTs } from '../../typescript/tests/validate';
 import { Types, mergeOutputs } from '@graphql-codegen/plugin-helpers';
 
 describe('TypeScript Mongo', () => {
